requirements.txt
asn1tools==0.146.6
atomicwrites==1.4.0
attrs==21.2.0
bitstring==3.1.6
bitstruct==8.11.1
colorama==0.4.4
cycler==0.10.0
diskcache==5.2.1
geographiclib==1.50
geopy==2.1.0
h5py==3.2.1
iniconfig==1.1.1
joblib==1.0.1
Keras==2.4.3
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.20.3
packaging==20.9
pandas==1.2.4
Pillow==8.2.0
pluggy==0.13.1
prompt-toolkit==3.0.18
py==1.10.0
pynmea2==1.15.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.3
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
selenium==3.141.0
six==1.16.0
sumolib==1.9.2
threadpoolctl==2.1.0
toml==0.10.2
traci==1.9.2
urllib3==1.26.5
wcwidth==0.2.5
